+++ optree/typing.py
@@ -282,7 +282,12 @@ def is_namedtuple_class(cls: type) -> bool:
         isinstance(cls, type)
         and issubclass(cls, tuple)
         and isinstance(getattr(cls, '_fields', None), tuple)
-        and all(isinstance(field, str) for field in cls._fields)  # type: ignore[attr-defined]
+        and all(
+            type(field) is str  # noqa: E721 # pylint: disable=unidiomatic-typecheck
+            for field in cls._fields  # type: ignore[attr-defined]
+        )
+        and callable(getattr(cls, '_make', None))
+        and callable(getattr(cls, '_asdict', None))
     )